Assume clipboard text from PoB is UTF-8 or ASCII
The Copy function previously put the copied text from Lua verbatim on the clipboard as `CF_TEXT` which if not annotated with a locale is assumed to be in the current "input locale".
This commit changes it to instead assume that it's either in UTF-8 (or ASCII, which is a subset of UTF-8), transcoding it to UTF-16LE for `CF_UNICODETEXT`.
This enables the ability for the Lua side to copy out UTF-8 text which it may have obtained from an external source like trade or another API.
A brief audit of the current Lua source code does not reveal any call site that should contain text in the local codepage as build XML is supposed to be UTF-8 and no file paths seem to be copied.