Add GetCloudProvider API function
This new `GetCloudProvider` function provides the ability to query
whether a directory is under a cloud provider sync root for a system
like OneDrive and whether that provider is active or inhibited.

This could be used to help diagnose user reports where the Lua code can
open but not read from files.