Cull geometry outside of viewport extents
Some views in the application can have a lot of geometry outside of the
viewports, particularly the passive skill tree.

We can save a lot of GPU work and emit fewer batch boundaries if we
conservatively cull out quads that are entirely outside of their
viewports by testing axis-aligned bounding boxes. This greatly speeds up
the skill tree when partially zoomed, which tends to be a common case.

This has benefits in reducing the amount of batch boundary cuts by not
having to consider the render states of unseen geometry, especially
texture changes.