Fix window positioning and maximization logic
The code to restore the window position and size from saved data did not
handle the case of being maximized on a non-primary monitor, nor did it
adjust the window position when the stored location was completely
outside of all active monitors.

This change addresses both these problems and removes some vestigal code
around display affinity and always-on-top logic which we don't use.