Restore screenshot functionality and Targa saving
With the migration to OpenGL ES2 from OpenGL 1.1 some parameters of
`glReadPixels` became more restrictive, only allowing two distinct pairs
of format/type; neither of which matched the format we needed to capture
screenshots.

This change reads out pixels in an always-supported format (RGBA) and
converts it to the desired RGB as it flips the image. Stubbed out
functionality to write Targa files was also reimplemented via stb_image.