Add C++ RAII scaffolding to handle API unwinding
In order to support C++ RAII classes in the functions exposed in the API
to Lua this change also introduces a new assertion `LExpect` functioning
like the existing `LAssert` but instead of directly transferring control
to Lua it pushes the message on the Lua stack and throws a C++ exception
to unwind the C++ call stack. For this scheme the convenience macros
`SG_LUA_CPP_FUN_BEGIN(Name)` and `SG_LUA_CPP_FUN_END()` wrap the
function entrypoint and perform the second phase of assertion processing
when needed.