feat: monkey-patch `io.open` to take UTF-8 paths
By squirreling away the original `io.open` function from Lua and
replacing the entry in the interpreter with our own wrapper that
interprets the path as UTF-8, we can now open files in locations that
cannot be represented in the user's codepage, as long as the path is
supplied as UTF-8.

Lua itself is very encoding agnostic, assuming that all strings are
soups of bytes and that paths are encoded in whatever the platform's
`FILE*` accepts. On Linux and macOS that's fairly friendly and in the
past dependent on the locale, but these days is de-facto standardized
on UTF-8.

On Windows the story is worse and narrow paths are assumed to be in the
user's current codepage, which can only express a fixed subset of
Unicode and which breaks if paths from the filesystem cannot be encoded
in the codepage. On Windows with NTFS the underlying filesystem is
UTF-16LE and supports all of Unicode, so it's good for consistency and
future work if we standardize on UTF-8, as Lua is very byte-oriented
when it comes to text and paths.