fix: harden UserPath against unusable folders
The user path location may both be inaccessible thanks to malfunctioning
cloud providers and due to not being representable in the user codepage
that is needed for interacting with the file system in a narrow way in
Lua.

If an error occurs, the Lua function returns nil and an approximation in
the user codepage of the path, suitable for use in an error message.