fix: make file searches case-insensitive
As we moved from WinAPI `FindFirstFile` to `std::filesystem` path tests
and globbing was accidentally made case-sensitive as we use direct
string comparison or regular expressions to match files.

We now always construct a regular expression for the glob and make the
regex match case-insensitive. The regex is rebuilt on every file test
but as the set of files is likely to be small this is of low impact;
caching the regex in the search object would spread RE2 into the header.