feat: expose and expect UTF-8 paths for Lua
Store the user path and base path as canonicalized std::filesystem paths
and expect the Lua side to be aware enough of UTF-8 to manipulate them
properly.

This together with the monkeypatched UTF-8 `io.open` and functions like
MakeDir and RemoveDir allows for use of a wider variety of installation
and user profile locations.