feat: new image "Art" type with image operations
Historically, Image handles represented an image resident in video
memory and were sourced from individual files on disk. As the game's
passive tree connectors are nested tightly in a single file, this means
that drawing such irregularly shaped would need richer geometric
primitives than what SimpleGraphic currently supports.

This change introduces an Art handle which represents an image in system
memory, making it available to image manipulation algorithms.

It can be used as an input to new texture loading functions in Image
intended to be able to handle the two kinds of connectors present in
the connector art:

* `LoadArtRectangle` - for slicing out the line connector;
* `LoadArtArcBand` - for slicing out the stacked orbital arc connectors.

Rectangles are defined by their corners where the upper left corner is
inclusive and the lower right corner is exclusive.

Arc bands are defined by the coordinate for the centre of the arc and by
inner and outer radii. The arc is always cut top-left of the centre.