feat: parallelise individual JPEG XL image loads
The reference JPEG XL library has the ability to parallelise the word of
decoding individual images. We use that here as a stopgap measure for
costly image loading at startup to try to cut down the time spent.

When/if we reintroduce async texture loading, this functionality may
need to be disabled to avoid oversubscription of CPU cores.

Note: this requires distribution of an additional `jxl_threads.dll` DLL.