fix: avoid deleting stack-local image
By wrapping an unique_ptr around a raw pointer argument in shader
loading we accidentally attempted to delete a stack-local image_c object
as the caller assumed that the callee would take ownership of the
internals but not the object itself.

This trap was removed by replacing the whole sequence of arguments with
an unique_ptr throughout.