feat: expose texture array layer for Lua drawing
The renderer is capable of binding and drawing individual texture layers
from 2D texture arrays. In the exporter and API we call these "stacked"
textures and as such the parameter is the "stack layer".

This improves on sprite sheets in that an individual image (quad) draw
can pick from any layer of its texture while retaining the ability to
have different wrapping/clamping modes and keeping any original mipmaps
from before stacking.

As this complicated the draw call parsing those are now parsed more
modularly based on the number of arguments to also handle stack layers
and also any possible future "mask" functionality we might want to draw
a subset of an image.

The immediate render interface now operates on vec2 values as the
argument counts were becoming ridiculous.