feat: allow image draws with stacked mask texture
DrawImage and DrawImageQuad now supports a final maskIdx that indicates
which texture in the stack to multiply the final output with; adding the
ability to mask out undesired parts of the image.

As it needs to be in the same stack, the mask image needs to have the
same size and image format as the colour image.