feat: use platform OpenGL if running on Wine
The Direct3D11 backend of ANGLE seems to have some trouble with the
stock Wine runtime where textures can fade into black like the big
class/ascendancy art in the middle of the tree at particular zoom levels
and also for error dialog boxes.

This can be mitigated by using the DXVK runtime with a Vulkan driver.
A cleaner more in-box workaround is to use the Desktop OpenGL backend of
ANGLE if we detect that the Windows version of SimpleGraphic runs on top
of Wine.