fix: detect Wine host, use D3D11 on macOS
MacOS doesn't support a cool enough desktop OpenGL version for us to
launch when using the Desktop OpenGL backend of ANGLE.

We now detect the presence of Wine and query it for the host platform
(analogous to `uname`).

On Wine macOS we select D3D11. This platform requires use of libraries
like D3DMetal/GPTK in Kegworks to emulate D3D11 enough.

On Wine Linux we select Desktop OpenGL as D3D11 requires use of DXVK
to function well enough.

On proper Windows we select D3D11 and for any other platforms we don't
indicate any preference for now.