feat: expose root paths as both UTF-8 and ACP
The standard encoding for all the top level paths (base, script, user)
in SimpleGraphic is UTF-8. For some purposes it's useful to know if the
path can be expressed in the user's active codepage as well, the updater
can use it to form absolute paths that the OS understands.

As the user path is fallible, we also expose a human-readable reason as
to why it could not be obtained.

All paths are made weakly canonical rather than fully canonical as the
full flavour requires paths to fully exist.